requirements.txt
pyrofork
tgcrypto
asyncio
aiofiles
pillow
pyleaves
telegraph
python-dotenv